+++ bin_unicorn/http/tls_client.h
@@ -15,12 +15,12 @@ typedef struct TLS_CLIENT_REQUEST_T_
 /// @param request Request details.
 /// @param buffer Buffer to write the response to.
 /// @param buffer_len The length of \p buffer.
-/// @return If greater than 0, the number of bytes written to \p buffer.
+/// @return 0 if successful.
 /// Otherwise, one of the following error codes:
 ///   * -1: Client timeout.
 ///   * -2: Failed to resolve DNS & open connection, or the tls_client_err callback was hit.
 ///   * -3: Failed to allocate client state.
 ///   * -4: Request was apparently successful, but no bytes were written to the buffer.
-int32_t https_get(TLS_CLIENT_REQUEST request, char *buffer, uint16_t buffer_len);
+int8_t https_get(TLS_CLIENT_REQUEST request, char *buffer, uint16_t buffer_len);
 
 #endif // HTTP_TLS_CLIENT_H_
